v2/requirements.txt
# Core dependencies for EasyCV
openai>=1.3.0
requests>=2.28.0
pathlib2>=2.3.0; python_version < "3.4"

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown2>=2.4.0

# Web generation
jinja2>=3.1.0

# Data handling
python-dateutil>=2.8.0

# Optional dependencies for enhanced features
# Style analysis
beautifulsoup4>=4.11.0

# Advanced markdown processing
markdown>=3.4.0
pymdown-extensions>=9.0

# Export enhancements
reportlab>=4.0.0  # For PDF generation (future feature)
weasyprint>=60.0  # For PDF from HTML (future feature)

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0 